+++ src/main/java/growup/spring/springserver/campaign/controller/CampaignController.java
@@ -1,14 +1,14 @@
-package growup.spring.springserver.campaign;
+package growup.spring.springserver.campaign.controller;
 
+import growup.spring.springserver.campaign.dto.CampaignResponseDto;
+import growup.spring.springserver.campaign.service.CampaignService;
 import growup.spring.springserver.global.common.CommonResponse;
-import growup.spring.springserver.login.dto.response.LoginResDto;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -25,10 +25,10 @@ public class CampaignController {
     @GetMapping("/getMyCampaigns")
     public ResponseEntity<CommonResponse<?>>getMyCampaigns(@AuthenticationPrincipal UserDetails userDetails) {
         log.info("Start getMyCampaigns API target is :"+userDetails.getUsername());
-        List<String> data = campaignService.getMyCampaigns(userDetails.getUsername());
+        List<CampaignResponseDto> data = campaignService.getMyCampaigns(userDetails.getUsername());
         log.info("End getMyCampaigns API target is :"+userDetails.getUsername());
         return new ResponseEntity<>(CommonResponse
-                .<List<String>>builder("success : load campaign name list")
+                .<List<CampaignResponseDto>>builder("success : load campaign name list")
                 .data(data)
                 .build(), HttpStatus.OK);
     }
